Apps Script Code
This is JavaScript code that I used in Apps Script. This code will take what I put in the connected spreadsheet and add it to the connected calendar. As you can see, I used the Spreadsheet ID and Google Calendar ID to connect the two. You can find documentation on google calendar and google sheets https://developers.google.com/apps-script/reference/calendar. I also have a program that deletes all events within a certain time range. This time range can be modified. For instance, futureDate takes the current time and adds (365  days * 24 hours * 60 minutes * 60 seconds* 1000 milliseconds), this makes the futureDate a year from whenever the code is ran. You can manipulate the numbers and even subtract time like was done in passDate.

I look forward to updating this code! I want to eventually optimize it so it automatically deletes events in the calendar whenever I delete the date in the spreadsheet.